Custom-Claims-Back-end/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0